chore(TDB): Setup of Error Handlers and Serialization

+++ src/todos/todos.service.ts
@@ -1,8 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { TodoRepository } from './todos.repository';
 import { EntityManager } from '@mikro-orm/core';
 import { Todo } from '../common/entities/todo.entity';
 import { CreateTodoDto, UpdateTodoDto } from './todos.dtos';
+import { TodoSerializer } from './todos.serializer';
 
 @Injectable()
 export class TodosService {
@@ -12,27 +17,63 @@ export class TodosService {
   ) {}
 
   async createTodo(createTodoDto: CreateTodoDto): Promise<Todo> {
-    const todo = this.todosRepository.create(createTodoDto);
-    await this.em.flush();
-    return todo;
+    try {
+      const todo = this.todosRepository.create(createTodoDto);
+      await this.em.flush();
+      return new TodoSerializer(todo).serialize() as unknown as Todo;
+    } catch (error) {
+      console.error('Error creating todo:', error);
+      throw new InternalServerErrorException('Failed to create a new task.');
+    }
   }
 
   async findAll(): Promise<Todo[]> {
-    return this.todosRepository.findAll();
+    try {
+      const todos = await this.todosRepository.findAll();
+      return todos.map(
+        (todo) => new TodoSerializer(todo).serialize() as unknown as Todo,
+      );
+    } catch (error) {
+      throw new InternalServerErrorException('Failed to fetch todos');
+    }
   }
 
   async findOne(id: string): Promise<Todo> {
-    return this.todosRepository.findOne(id);
+    try {
+      const todo = await this.todosRepository.findOne(id);
+      if (!todo) {
+        throw new NotFoundException('Todo not found');
+      }
+      return new TodoSerializer(todo).serialize() as unknown as Todo;
+    } catch (error) {
+      if (error instanceof NotFoundException) {
+        throw error;
+      } else {
+        throw new InternalServerErrorException('Failed to fetch todo');
+      }
+    }
   }
 
   async updateTodo(id: string, updateTodoDto: UpdateTodoDto): Promise<Todo> {
-    const todo = this.todosRepository.update(id, updateTodoDto);
-    await this.em.flush();
-    return todo;
+    try {
+      const todo = this.todosRepository.update(id, updateTodoDto);
+      await this.em.flush();
+      return new TodoSerializer(todo).serialize() as unknown as Todo;
+    } catch (error) {
+      throw new InternalServerErrorException('Failed to update todo');
+    }
   }
 
   async removeTodo(id: string): Promise<void> {
-    this.todosRepository.remove(id);
-    await this.em.flush();
+    const todo = await this.todosRepository.findOne(id);
+    if (!todo) {
+      throw new NotFoundException(`Todo with ID ${id} not found`);
+    }
+    try {
+      this.todosRepository.remove(id);
+      await this.em.flush();
+    } catch (error) {
+      throw new InternalServerErrorException('Failed to remove todo');
+    }
   }
 }
